Updated file structure
no refs.
- added new controller/routes and template for Labs -> Members
- moved all members related settings to new page
- removed members related settings from Labs
