+++ app/router.js
@@ -45,6 +45,7 @@ Router.map(function () {
 
     this.route('settings.general', {path: '/settings/general'});
     this.route('settings.labs', {path: '/settings/labs'});
+    this.route('settings.labs.members', {path: '/settings/labs/members'});
     this.route('settings.code-injection', {path: '/settings/code-injection'});
     this.route('settings.design', {path: '/settings/design'}, function () {
         this.route('uploadtheme');
